Add files via upload
In this project , EDA is performed on amazon_alexa.tsv dataset available on Kaggle , in which the reviews are there of users on different variations of Alexa they are using and a target column (feedback , 1 - positive , 0- negative). After EDA , Data Preprocessing and then training 3 models (Random Forest Classifier , XGBoost and Decision tree classifier) is done and model is saved.